requirements.txt
streamlit>=1.35.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
langgraph-checkpoint-sqlite>=2.0.0
langsmith>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
